requirements.txt
click==8.0.3
colorama==0.4.4
requests==2.27.1
python-dotenv==0.19.2
